remove the abs on the area calculus

+++ python-abc/task_01_duck_typing.py
@@ -28,7 +28,7 @@ class Circle(Shape):
     Circle
     """
 
-    def __init__(self, radius: int):
+    def __init__(self, radius: float):
         self.radius = radius
 
     def area(self):
